+++ src/Calculation/CalculationInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Northwestern\SysDev\DynamicForms\Calculation;
+
+interface CalculationInterface
+{
+    public function __invoke(array $submissionValues): mixed;
+}