+++ src/Calculation/JSONCalculation.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace Northwestern\SysDev\DynamicForms\Conditional;
+namespace Northwestern\SysDev\DynamicForms\Calculation;
 
 use JWadhams\JsonLogic;
 use Northwestern\SysDev\DynamicForms\JSONLogic\JsonLogicHelpers;
 
-class JSONConditional implements ConditionalInterface
+class JSONCalculation implements CalculationInterface
 {
     public function __construct(array $jsonLogic)
     {
         $this->jsonLogic = JsonLogicHelpers::convertDataVars($jsonLogic);
     }
 
-    public function __invoke(array $submissionValues): bool
+    public function __invoke(array $submissionValues): mixed
     {
         return JsonLogic::apply($this->jsonLogic, $submissionValues);
     }