+++ src/Components/ComponentInterface.php
@@ -3,6 +3,7 @@
 namespace Northwestern\SysDev\DynamicForms\Components;
 
 use Illuminate\Contracts\Support\MessageBag;
+use Northwestern\SysDev\DynamicForms\Calculation\CalculationInterface;
 use Northwestern\SysDev\DynamicForms\Conditional\ConditionalInterface;
 
 interface ComponentInterface
@@ -27,6 +28,7 @@ interface ComponentInterface
         ?array $conditional,
         ?string $customConditional,
         string $case,
+        null|array|string $calculateValue,
         array $additional,
     );
 
@@ -75,6 +77,16 @@ interface ComponentInterface
     public function canValidate(): bool;
 
     /**
+     * Whether this component's value is calculated or not.
+     *
+     * A calculated field will have its calculated value replace a value submitted
+     * by the form. The server-side calculated value is always returned.
+     *
+     * @return bool
+     */
+    public function isCalculated(): bool;
+
+    /**
      * Run validations for the component.
      */
     public function validate(): MessageBag;
@@ -99,6 +111,11 @@ interface ComponentInterface
      * Returns an invokable Conditional instance.
      */
     public function conditional(): ?ConditionalInterface;
+
+    /**
+     * Returns an invokable Calculation instance.
+     */
+    public function calculation(): ?CalculationInterface;
 
     /**
      * List of transformations to apply to the value.